Fix type annotation of center property in DataSet

+++ pyvista/core/dataset.py
@@ -1774,7 +1774,7 @@ class DataSet(DataSetFilters, DataObject):
         return self.GetLength()
 
     @property
-    def center(self) -> List[int]:  # numpydoc ignore=RT01
+    def center(self) -> List[float]:  # numpydoc ignore=RT01
         """Return the center of the bounding box.
 
         Returns
